Convert Reference#reference_references to a service object
